Disable LEDs via configuration and not at compile time
Start passing --enable-leds to configure again and disable LEDs by
setting ui.leds.enable to 0.  This will ensure that there are no wakeups
with default installations on laptops and people who like the feature
can enable it in their local configuration.

Thanks to ﺄﺤﻣﺩ ﺎﻠﻤﺤﻣﻭﺪﻳ (Ahmed El-Mahmoudy) for the suggestion and for
testing this.