Remove "-b -B debian/patched/" passed to patch while patching
This means that the original files aren't backed up anymore in the
debian/patched/ directory when "debian/rules patch" is invoked.  This is
not required with the new git-diff patch system.  Also if multiple
patches modify a file, this was storing the previously patched file as
opposed to the real original (which is misleading).